mc-9245 Fix security access around DataClass and DataElements
When trying to get objects inside a DataClass we weren't checking the relationship between the DataClass and the DataModel, as we assumed the security has checked the user's acccess to the datamodel.
This meant using a DataClass id with a dataModel id which you could read allowed you to access dataclasses and dataelements inside this dataclass, although not the dataclass directly.

Therefore add searching to include the datamodel id in the queries and in the case of DataElements check the Dataclass exists inside the DataModel first