Alter methods to remove facets from catalogueitems in the service
* Custom code the SQL to remove the entry from the join table

The issue is that if the collection is empty the ORM code just does a blanket removal from the join table rather than doing a removal by facet id. This results in the join table having all facet links removed. The custom code makes sure we delete by id.

However we need to make sure the removeFrom methods are never used and the service method is always used otherwise we have this issue.