gh-76 Fix the NPE on trying to find a CatalogueItemService
THe grailsApplication was being set to null when plugins loaded into the mdm-application-build.
Quite why its null under this situation and not when under functional test i have no idea, however i have replaced the autowired annotation which fixes it.

While doing this I have just done a small amount of tidy up the Profile controller to use respond which sets and handles json responses whereas render does not do this properly.
Also removed "Object" references in the gson views as we know they are always of a specific supertype
Finally moved the unused logic into the service where it belongs

Added a functional test for both used and unused endpoints just to double down on the fact they do actually work