mc-9528 Add flyway scripts to handle the version link type changes
We recoded model version to fork version and then removed the "superseded by model version" as it would never be used.
However we forgot about migrated from MDC instances which would be using this.
These scripts update the version link types to their correct types and then also swop the deprecated types to their un-deprecated types.

We can now remove the deprecated types as this script will prevent their existence in the database