Rules support added,
bug in multifacetItemAware(line 44) aware when forking/copying a rule , it will copy the multiFacetItemAware object along with the rule but that object will have an empty id. it then assigns the empty id to the MultiFacetAware object and causes a validation error. Not sure if this is related to it not assigning a new MFA an id  when copying or something else that will be repeated with other MDA implimentations.