gh-177 Fix the underlying issue around finalising and BTs
When a model was finalised and a BT already existed it wasnt updating the BT finalised state properly.
Which means that when the DC and other content BTs were rendered or checked the BT still thought it was unfinalised.
This fix makes sure that changing the state of the model's finalise field updates the BT finalise field.
The BTs are pre-created in the checkModelAssociations method which then results in the checkTree method being called which only checks that the last Breadcrumb matches the expected Breadcrumb, therefore the fact the DM is now finalised is missed.
If the model is imported with finalised in the import then this is not an issue as the DM BT will be marked as finalised.