mc-9723 Recode and reorganise the GroupBasedSecurityPolicyManagerService and GroupBasedUserSecurityPolicyManager
* Split out the policy information into a new UserSecurityPolicy
  This can be better locked and getter methods controlled by checking the lock status
* Add a lock, unlock and destroy to the UserSecurityPolicy.
  This will signify if the policy is locked for changes or has been destroyed
* Split out actions on the new UserSecurityPolicy into a UserSecurityPolicyService.
  This makes the GroupBasedSecurityPolicyManagerService cleaner as all the building of the policy is now done inside the UserSecurityPolicyService
* Have the UserSecurityPolicyManagerInterceptor hold if the current policy is locked

This is yet to be tested but in theory the locking should act to prevent access whilst policies are being updated.
This will help when we rebuild a policy and its access is revoked.
The Underlying Cache is a ConcurrentHashMap which has no difference at the atomic level between put and putIfAbsent, which may have changed in later versions where putIfAbsent is supposed to be immediate.
However putIfAbsent will not update the cache if its there already so we must keep using put.