gh-183 Update the security to correctly handle removing public or authenticated access
When removing access at the folder level we were blanket removing all access below, ignoring any access provided by the content.

The code has been updated to only build up access roles if adding the readableByXxxx.
As we are expanding access into sets adding roles which already exist doesnt matter

If we are removing readableByXxx we now search all policies which currently have access to that resource.
Then we rebuild those policies completely, this is the only way to ensure the access levels are correctly maintained.

The code has been reviewed for were other access is removed.
However in other locations this is done by including the UserGroup as part of the role, or by completely rebuilding already.