Make sure the models are saved appropriately during the first pass of copying a folder
Rather than do all at the end which will cause issues, do each at the appropriate time.
We avoid the proxy/object not loaded issue by retrieving the working model and the working folder from the database before we start.
This way when the session is cleared the required components are reloaded back into the session