Improve branching/copying
* Perform a quick check on whether the folder contains any models for that service. count is faster than find.
* Add some additional timing logging
* Add the ability to cache facet data into the CopyInformation class
* Preload the modelitem class's facets into a CopyInformation object and pass that to the copy call
* Load the MIs using a custom SQL call with classifiers joined.

This means we only call the 3 facet SQL once per group of MIs rather than 3 times per MI when we copy.
We handle the case where the cache is not available as it wont be when we use the copy endpoint at that exact level.
However this will speed up copying as its directly used at the copy method not the branch method.

The loading of MIs reduces the separate calls for each MI and also means we dont have to call the classifier service for the classifiers as we already have them loaded