Remove the last deprecated methods and rename relevant fields from lucene to hibernatesearch
in the migration to HS6 some of the methods which contain the word lucene were missed.
This refactors everything which is relevant to replace with hibernateSearch as for the most part we're backend agnostic.

We do have some minor lucene specific code but we could recode this to elasticsearch if we wanted to change.