Store loaded profile jsons using a unique key per profile provider
Currently we just cache the profiles by the namespace, however this doesnt allow multiple profiles to use the same namespace.
This is a problem in a number of plugins which use 1 MD NS for all their MD but need to provide different profiles for CIs.
Also this will allow different versions of the same profile to exist.
This won't break anything as it just changes how we cache the profile jsons.
The key is made up from the NS.Name.Version which is what is used to "Search" for  a profile provider.