mc-9803 Improve the login process
* Use grails async to push the building of the actual user policy off to an async process
* Set the async generated policy to a lock of 60s to allow the policy time to build
* Make sure the Usergroups in the Policy are a normal HashSet not a GORM persistent set
* Only use async build in the authentication process, however leave the methods open for other locations to make use.

Return a simple policy which is locked with no access.
This allows users to start using the system while the policy itself is built.
The lock will ensure any calls made will hold until the lock is released.
Remove persistent set usergroups which wont work as their session is not the same as the one we're trying to process the build inside