Add AsyncJobFuture to properly manage the task state.
Wrapping class to combine the actual future and the task,
allowing us to know the current step state of the task and make sure we wait/respond appropriately.

This arises because the main task cannot be interrupted once started,
therefore we have to wait for it to finish then
rollback the transaction.
However if we dont acknowledge that wait and rollback the task could be asumed to be finished and "isDone" before it actually is