Add import metadata to diffs
This can be extracted in the same manner as we do for copying, and then added to the diffcache or diff.
We do have to add this at the importing element diff level.
For the diff this may look odd, however for the merge diff it will just appear as a path and therefore should work and look fine